Modified WITs to modularity
